test(app): complete #4224 naval panel family densify AC7
Extract combine/autoclose tables and part1/mockup pins into naval_panel_*
siblings; consolidate part1 pumps; merge mockup fidelity into part3.
Core naval_units_panel_* family ~1939 LOC vs 3217 baseline (~40% reduction).

Refs #4224
